refactor(anytype): split importer into model + content modules
Separate the shared types into model.ts (the export shape and Parsed* values) and
the block-tree -> HTML rendering into content.ts (extractContent, inline marks,
code blocks), mirroring the Notion importer's model.ts / converter.ts layout.

importer.ts shrinks from ~780 to ~295 lines (now just orchestration, zip reading,
object predicates, parseObject and the note-tree builder) and no longer imports
escapeHtml or the mime_type helpers. collection.ts imports its types from model.ts,
so the former type-only cycle with importer.ts is gone; the four modules are
acyclic (importer/content/collection -> model).

Relocate the rendering tests to content.spec.ts with their assertions unchanged;
importer.spec.ts keeps isPage / parseObject / dates. Verified with tsc --build.
